+++ starsky/starsky/clientapp/src/hooks/realtime/use-sockets.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, SetStateAction, useEffect, useRef, useState } from "react";
 import { DifferenceInDate } from "../../shared/date";
 import useInterval from "../use-interval";
-import { IsClientSideFeatureDisabled } from "./internal/is-client-side-feature-disabled.ts1";
+import { IsClientSideFeatureDisabled } from "./internal/is-client-side-feature-disabled.ts";
 import WebSocketService from "./websocket-service";
 import WsCurrentStart, { NewWebSocketService } from "./ws-current-start";
