+++ starsky/starsky/clientapp/src/hooks/use-location/internal/navigate-fn.ts
